# ent/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ent Schema API",
    "description": "This is an auto generated API description made out of an Ent schema definition",
    "version": "0.1.0"
  },
  "paths": {
    "/orders": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "List Orders",
        "description": "List Orders.",
        "operationId": "listOrder",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "what page to render",
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "itemsPerPage",
            "in": "query",
            "description": "item count to render per page",
            "schema": {
              "type": "integer",
              "maximum": 255,
              "minimum": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "result Order list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderList"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "tags": [
          "Order"
        ],
        "summary": "Create a new Order",
        "description": "Creates a new Order and persists it to storage.",
        "operationId": "createOrder",
        "requestBody": {
          "description": "Order to create",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_type_example_1": {
                    "type": "string",
                    "enum": [
                      "PHYSICAL",
                      "DIGITAL",
                      "SERVICE"
                    ]
                  },
                  "item_type_example_2": {
                    "type": "integer"
                  },
                  "item_type_example_3": {
                    "type": "string",
                    "enum": [
                      "PHYSICAL",
                      "DIGITAL",
                      "SERVICE"
                    ]
                  }
                },
                "required": [
                  "item_type_example_1",
                  "item_type_example_2",
                  "item_type_example_3"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderCreate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "tags": [
          "Order"
        ],
        "summary": "Find a Order by ID",
        "description": "Finds the Order with the requested ID and returns it.",
        "operationId": "readOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Order",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order with requested ID was found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "delete": {
        "tags": [
          "Order"
        ],
        "summary": "Deletes a Order by ID",
        "description": "Deletes the Order with the requested ID.",
        "operationId": "deleteOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Order",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Order with requested ID was deleted"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "patch": {
        "tags": [
          "Order"
        ],
        "summary": "Updates a Order",
        "description": "Updates a Order and persists changes to storage.",
        "operationId": "updateOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the Order",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Order properties to update",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "item_type_example_1": {
                    "type": "string",
                    "enum": [
                      "PHYSICAL",
                      "DIGITAL",
                      "SERVICE"
                    ]
                  },
                  "item_type_example_2": {
                    "type": "integer"
                  },
                  "item_type_example_3": {
                    "type": "string",
                    "enum": [
                      "PHYSICAL",
                      "DIGITAL",
                      "SERVICE"
                    ]
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Order updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderUpdate"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "409": {
            "$ref": "#/components/responses/409"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type_example_1": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          },
          "item_type_example_2": {
            "type": "integer"
          },
          "item_type_example_3": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          }
        },
        "required": [
          "id",
          "item_type_example_1",
          "item_type_example_2",
          "item_type_example_3"
        ]
      },
      "OrderCreate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type_example_1": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          },
          "item_type_example_2": {
            "type": "integer"
          },
          "item_type_example_3": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          }
        },
        "required": [
          "id",
          "item_type_example_1",
          "item_type_example_2",
          "item_type_example_3"
        ]
      },
      "OrderList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type_example_1": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          },
          "item_type_example_2": {
            "type": "integer"
          },
          "item_type_example_3": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          }
        },
        "required": [
          "id",
          "item_type_example_1",
          "item_type_example_2",
          "item_type_example_3"
        ]
      },
      "OrderRead": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type_example_1": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          },
          "item_type_example_2": {
            "type": "integer"
          },
          "item_type_example_3": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          }
        },
        "required": [
          "id",
          "item_type_example_1",
          "item_type_example_2",
          "item_type_example_3"
        ]
      },
      "OrderUpdate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "item_type_example_1": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          },
          "item_type_example_2": {
            "type": "integer"
          },
          "item_type_example_3": {
            "type": "string",
            "enum": [
              "PHYSICAL",
              "DIGITAL",
              "SERVICE"
            ]
          }
        },
        "required": [
          "id",
          "item_type_example_1",
          "item_type_example_2",
          "item_type_example_3"
        ]
      }
    },
    "responses": {
      "400": {
        "description": "invalid input, data invalid",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "403": {
        "description": "insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "404": {
        "description": "resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "409": {
        "description": "conflicting resources",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      },
      "500": {
        "description": "unexpected error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer"
                },
                "status": {
                  "type": "string"
                },
                "errors": {}
              },
              "required": [
                "code",
                "status"
              ]
            }
          }
        }
      }
    }
  }
}